+++ docs/CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Security
 
 ### Added
+* Added a script that automates the installation process and can be run
+  with just one single command. See
+  [here](getting-started.html#one-step-installer) for more information.
 
 ### Changed